Rename tool to FetchPullRequestData
