+++ src/FetchPullRequestData/CommandLine.cs
@@ -1,6 +1,6 @@
 ﻿using CommandLine;
 
-namespace ApiTool
+namespace FetchPullRequestData
 {
     class CommandLineOptions
     {
