+++ src/FetchPullRequestData/FetchPullRequestData.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.28307.757
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ApiTool", "ApiTool.csproj", "{18107F5C-D68B-48A5-91F3-2114C802EAA2}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "FetchPullRequestData", "FetchPullRequestData.csproj", "{18107F5C-D68B-48A5-91F3-2114C802EAA2}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
